Added more explanation about build systems and reduced conflictiing relationships
I never understood why Breaks: was set this way, there is no real
conflict of the -dkms package and even if the user attempts to build and
install something with m-a locally then a possibly conflicting dkms
package will be removed. On the other hand, there might be valid
usecases to have dkms thing locally installed and use -source to build
the package for another system (second laptop, for example).