d/patches: Fix the linux-5.17 deprecations patch
The patch was enough to get things to build, but it accidentally
inverted the direction of the copy of a few addresses.

This was causing traces when the device address was set in the driver:

```
    wl 0000:03:00.0 wlp3s0: Current addr:  NN NN NN NN NN NN 00 00 00 00 00 00 00 00 00
    wl 0000:03:00.0 wlp3s0: Expected addr: 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
    ------------[ cut here ]------------
    netdevice: wlp3s0: Incorrect netdev->dev_addr
    WARNING: CPU: 3 PID: 1178 at net/core/dev_addr_lists.c:517 dev_addr_check.cold+0x43/0x7d
    (...)
```

(where NN is the device MAC)

Apparently this might be enough in some cards to make it impossible to
connect to a network. See bugs below.

See: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1011529
See: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1016426